Some more fixes + the start of the release page.
