--- Assets/Scripts/GameManager/GameManager.UI.cs
@@ -1,8 +1,0 @@
-public partial class GameManager
-{
-    public string CompanyName => _companyName;
-    public int UnAssignedEmployees => _unAssignedEmployees;
-    public int TotalEmployees => _totalEmployees;
-    public double Budget => _budget;
-    public double CashFlow => _cashFlow;
-}
