--- Assets/Scripts/GameManager/GameManager.Variables.cs
@@ -1,9 +1,0 @@
-public partial class GameManager
-{
-    private string _companyName;
-    private int _unAssignedEmployees;
-    private int _assignedEmployees;
-    private int _totalEmployees;
-    private double _budget;
-    private double _cashFlow;
-}
