+++ Assets/Scripts/GameManagerPartials/GameManager.UI.cs
@@ -1,0 +1,11 @@
+namespace GameManagerPartials
+{
+    public partial class GameManager
+    {
+        public string CompanyName => _companyName;
+        public int UnAssignedEmployees => _unAssignedEmployees;
+        public int TotalEmployees => _totalEmployees;
+        public double Budget => _budget;
+        public double CashFlow => _cashFlow;
+    }
+}
