+++ Assets/Scripts/GameManagerPartials/GameManager.Variables.cs
@@ -1,0 +1,12 @@
+namespace GameManagerPartials
+{
+    public partial class GameManager
+    {
+        private string _companyName;
+        private int _unAssignedEmployees;
+        private int _assignedEmployees;
+        private int _totalEmployees;
+        private double _budget;
+        private double _cashFlow;
+    }
+}
