+++ Assets/Scripts/GameManagerPartials/GameManager.Variables.cs.meta
