+++ Assets/Scripts/UI/ReleasePage.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 60be41f0973603641b58cc9121ac9590
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
