+++ Assets/Scripts/UI/ReleasePage/ReleasePageMainHandler.cs
@@ -1,0 +1,14 @@
+using UnityEngine;
+
+public class ReleasePageMainHandler : MonoBehaviour
+{
+    // Start is called before the first frame update
+    void Start()
+    {
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+    }
+}
