+++ Assets/Scripts/UI/ReleasePage/ReleasePageMainHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4b462328cf216044682f8b9a07595341
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
